fix a bug caused by sql

+++ wxshop-main/src/main/java/com/github/wxshop/controller/OrderController.java
@@ -54,7 +54,7 @@ public class OrderController {
     }
 
     @RequestMapping(value = "/order/{id}", method = {RequestMethod.POST, RequestMethod.PATCH})
-    public Response<OrderResponse> updateOrder(@PathVariable("id") Integer id, @RequestBody Order order) {
+    public Response<OrderResponse> updateOrder(@PathVariable("id") long id, @RequestBody Order order) {
         if (order.getExpressCompany() != null) {
             return Response.of(orderService.updateExpressInformation(order, UserContext.getCurrentUser().getId()));
         } else {
